feat(scoring): Phase γ history schema int → float + minmax align (#18)
`ScoreHistory.post_count` schema 를 int → float 로 마이그 (β2/β4 share-weighted 정합).
`_read_count` 가 `float()` cast 로 기존 int json 도 자연 호환 (read-cast 정책 — locked).
영역:
- `ScoreHistory.get_total_post_count` / `get_post_count_history` return type → float /
  list[float]. `sum(..., 0.0)` 로 빈 cluster 도 float 보장.
- `ScoreHistory.update(post_count: int → float)` / `WeeklyScoreHistory.update_weekly` 동.
- `ClusterScoringContext.post_count_total: int → float` (β2 mixed precision round 제거).

`_build_contexts` 의 `round(a.post_count_today)` 제거 → `history.get_total + share-sum`
직접 합 (β1 effective_item_count_today ↔ history total ↔ ctx.post_count_total 모두
float 단위, minmax 분자/분모 align).

`_decide_clusters` 는 ClusterDecision contract 호환 위해 `int(ctx.post_count_total)` /
`round(ctx.post_count_today)` 1회 cast (presentation 시점 단일 round). Output contract
(`TrendClusterSummary.post_count_*: int`) 는 그대로 — 별도 PR 영역 (StarRocks DDL 영향).

`score_and_export` 의 `post_count_int = round(...)` 제거 → `ctx.post_count_today` float
그대로 history/weekly 적재. 누적 시점 round 손실 0.

신규 4 핀 (`tests/unit/test_score_history_float_migration.py`):
- 기존 int json read → float cast
- update float → save → reload round-trip 정확
- empty cluster get_total → 0.0 (float 일관)
- WeeklyScoreHistory update_weekly float 호환

763 passed (+4 신규), regression 0.

+++ src/scoring/cluster_context.py
@@ -41,7 +41,7 @@ class ClusterScoringContext(BaseModel):
     momentum_new_account_ratio: float     # 0~1.
 
     # --- 라이프사이클 / 방향성 / maturity 판정용 ---
-    post_count_total: int                 # 누적 (직전일까지 + 오늘). Phase γ TODO: history schema 마이그 후 float
+    post_count_total: float               # Phase γ: history schema 마이그 후 float (β2 mixed precision round 제거)
     post_count_today: float               # Phase β2: share-weighted fan-out 으로 fractional
     avg_engagement_rate: float
     # 전일 / 지난 주 같은 비교 기준은 orchestrator 가 별도 dict 로 넘긴다.
